requirements.txt
django
psycopg2-binary
django-leaflet
requests
geopy